requirements.txt
matplotlib>=3.8
numpy>=1.26
pandas>=2.0
Pillow>=10.0
pydantic>=2.10
pydantic_ai>=0.2.4
pytorch_lightning>=2.3
reportlab>=4.0
seaborn>=0.13
torch>=2.3
torchvision>=0.18
tqdm>=4.66
httpx>=0.27
medmnist>=3.0
scipy>=1.11
